wip: update chat style
